Organize index operations into discrete files
