+++ src/execution_plan/ops/op_delete.c
@@ -84,7 +84,7 @@ void _DeleteEntities(OpDelete *op) {
         if (op->g->index_ctr > 0) {
           // Build a triemap with index IDs as keys and Vectors containing
           // all matching nodes as values
-          TrieMap *delete_map = Indices_BuildDeletionMap(op->ctx, op->g, op->graph_name, op->deleted_nodes, NULL, dupFreeNodeIDsCount);
+          TrieMap *delete_map = Indices_BuildModificationMap(op->ctx, op->g, op->graph_name, op->deleted_nodes, NULL, dupFreeNodeIDsCount);
           if (delete_map) {
             // Traverse the triemap and delete all described nodes from indices
             Indices_DeleteNodes(op->ctx, op->g, op->graph_name, delete_map);
@@ -94,7 +94,7 @@ void _DeleteEntities(OpDelete *op) {
           EntityID *migrations = _Graph_EnqueueMigrations(op->g->nodes, op->deleted_nodes, dupFreeNodeIDsCount, &migration_count);
           // Build a triemap with index IDs as keys and Vectors containing
           // all matching nodes and their replacement IDs as values
-          TrieMap *replacement_map = Indices_BuildDeletionMap(op->ctx, op->g, op->graph_name, migrations, op->deleted_nodes, migration_count);
+          TrieMap *replacement_map = Indices_BuildModificationMap(op->ctx, op->g, op->graph_name, migrations, op->deleted_nodes, migration_count);
           if (replacement_map) {
             Indices_UpdateNodeIDs(op->ctx, op->g, op->graph_name, replacement_map);
           }
