+++ src/graph/graph.h
@@ -83,7 +83,7 @@ void Graph_ConnectNodes (
         DataBlockIterator **edges   // Pointer to edge iterator.
 );
 
-// Populate an array of label iDs associated with a given node
+// Populate an array of label IDs associated with a given node
 // and return the label count.
 size_t Graph_GetNodeLabels(const Graph *g, NodeID id, size_t **labels);
