+++ src/index/index_iter.c
@@ -1,0 +1,28 @@
+#include "index_iter.h"
+
+/* Generate an iterator with no lower or upper bound. */
+IndexIter* IndexIter_Create(Index *idx, SIType type) {
+  skiplist *sl = type == T_STRING ? idx->string_sl : idx->numeric_sl;
+  return skiplistIterateAll(sl);
+}
+
+/* Apply a filter to an iterator, modifying the appropriate bound if
+ * it narrows the iterator range.
+ * Returns 1 if the filter was a comparison type that can be translated into a bound
+ * (effectively, any type but '!='), which indicates that it is now redundant. */
+bool IndexIter_ApplyBound(IndexIter *iter, SIValue *bound, int op) {
+  return skiplistIter_UpdateBound(iter, bound, op);
+}
+
+NodeID* IndexIter_Next(IndexIter *iter) {
+  return skiplistIterator_Next(iter);
+}
+
+void IndexIter_Reset(IndexIter *iter) {
+  skiplistIterate_Reset(iter);
+}
+
+void IndexIter_Free(IndexIter *iter) {
+  skiplistIterate_Free(iter);
+}
+
